+++ socialhome/federate/utils/generic.py
@@ -1,5 +1,6 @@
 import datetime
 import logging
+import pickle
 
 import django_rq
 from django.conf import settings
@@ -7,6 +8,7 @@ from django.contrib.sites.models import Site
 from django.http import HttpRequest
 from django.utils.timezone import now
 from dynamic_preferences.registries import global_preferences_registry
+from federation.types import RequestType
 
 from socialhome import __version__ as version
 from socialhome.content.enums import ContentType
@@ -28,8 +30,7 @@ def get_nodeinfo2_data():
             "software": "socialhome",
             "version": version,
         },
-        # TODO fix when relay is configurable
-        "relay": "all",
+        "relay": settings.SOCIALHOME_RELAY_SCOPE,
         "openRegistrations": settings.ACCOUNT_ALLOW_REGISTRATION,
     }
     if settings.SOCIALHOME_STATISTICS:
@@ -51,17 +52,35 @@ def get_nodeinfo2_data():
     return data
 
 
-def queue_payload(request: HttpRequest, uuid: str=None):
+def queue_payload(request: HttpRequest, uuid: str = None):
     """
     Queue payload for processing.
     """
     from socialhome.federate.tasks import receive_task  # Circulars
     try:
-        payload = request.body
+        # Create a simpler request object we can push to RQ
+        headers = {}
+        for key, value in request.META.items():
+            key = key.replace('HTTP_', '').lower().replace('_', '-').capitalize()
+            try:
+                pickle.dumps(value)
+            except Exception:
+                pass
+            else:
+                headers[key] = value
+                # Include also a lowercase version for compatibility with signature verification module
+                headers[key.lower()] = value
+        _request = RequestType(
+            body=request.body,
+            headers=headers,
+            method=request.method,
+            url=request.build_absolute_uri(),
+        )
         preferences = global_preferences_registry.manager()
         if preferences["admin__log_all_receive_payloads"]:
-            logger.debug("get_payload_from_request - Payload: %s", payload)
-        django_rq.enqueue(receive_task, payload, uuid=uuid)
+            logger.debug("queue_payload - Request: %s", _request)
+        django_rq.enqueue(receive_task, _request, uuid=uuid)
         return True
     except Exception:
+        logger.exception('Failed to enqueue payload')
         return False