+++ socialhome/frontend/components/streams/stamped_elements/ProfileStampedElement.vue
@@ -51,7 +51,7 @@
             <div class="mt-1">
                 <b-button
                     variant="secondary"
-                    href="#"
+                    :href="urls.contactsFollowers"
                     :title="translations.followers"
                     :aria-label="translations.followers"
                 >
@@ -98,7 +98,7 @@
         <div class="d-inline-block ml-3 align-center stamped-profile-info">
             <h1>{{ displayName }}</h1>
             <h3>
-                <cite :title="translations.userHandle">{{ profile.handle.length ? profile.handle : profile.fid }}</cite>
+                <cite :title="translations.userHandle">{{ profileHandle }}</cite>
             </h3>
         </div>
         <div class="text-center">
@@ -136,6 +136,9 @@ export default Vue.component("profile-stamped-element", {
         profile() {
             return this.$store.state.application.profile
         },
+        profileHandle() {
+            return this.profile.handle ? this.profile.handle : this.profile.fid
+        },
         showProfileButtons() {
             return this.$store.state.application.isUserAuthenticated &&
                 this.profile.id === this.$store.state.application.currentBrowsingProfileId
@@ -159,6 +162,7 @@ export default Vue.component("profile-stamped-element", {
         urls() {
             return {
                 contactsFollowed: Urls["users:contacts-followed"](),
+                contactsFollowers: Urls["users:contacts-followers"](),
                 organizeProfileUrl: Urls["users:profile-organize"](),
                 pictureUpdate: Urls["users:picture-update"](),
                 pinnedContent: Urls["users:profile-detail"]({uuid: this.profile.uuid}),