+++ socialhome/users/tests/test_views.py
@@ -12,7 +12,7 @@ from socialhome.streams.enums import StreamType
 from socialhome.tests.utils import SocialhomeTestCase
 from socialhome.users.models import User, Profile
 from socialhome.users.serializers import ProfileSerializer
-from socialhome.users.tables import FollowedTable
+from socialhome.users.tables import ContactTable
 from socialhome.users.tests.factories import (
     UserFactory, AdminUserFactory, ProfileFactory, PublicUserFactory, PublicProfileFactory)
 from socialhome.users.views import (
@@ -489,7 +489,33 @@ class TestContactsFollowedView(SocialhomeTestCase):
     def test_contains_table_object(self):
         with self.login(self.user):
             self.get("users:contacts-followed")
-        self.assertTrue(isinstance(self.context["followed_table"], FollowedTable))
+        self.assertTrue(isinstance(self.context["followed_table"], ContactTable))
+        self.assertEqual(len(self.context["followed_table"].data), 1)
+        self.assertContext("profile", self.user.profile)
+
+
+class TestContactsFollowersView(SocialhomeTestCase):
+    @classmethod
+    def setUpTestData(cls):
+        super().setUpTestData()
+        cls.user = UserFactory()
+        cls.profile = ProfileFactory()
+        cls.profile.following.add(cls.user.profile)
+
+    def test_login_required(self):
+        # Not logged in, redirects to login
+        self.get("users:contacts-followers")
+        self.response_302()
+        # Logged in
+        with self.login(self.user):
+            self.get("users:contacts-followers")
+        self.response_200()
+
+    def test_contains_table_object(self):
+        with self.login(self.user):
+            self.get("users:contacts-followers")
+        self.assertTrue(isinstance(self.context["followers_table"], ContactTable))
+        self.assertEqual(len(self.context["followers_table"].data), 1)
         self.assertContext("profile", self.user.profile)